Use pre-defined widths for tablets when sheet params are WRAP_CONTENT
This introduces some proper width handling on tablets. Forcing all sheets to be the same width as the parent can be a bad UI experience, especially for short lines of text.

This is based on what the system does for the width of the notification panel, with custom sizes for sw600dp and sw720dp screen densities.