Make bottomsheet work with YoutubePlayer from the Youtube SDK
The player pauses if something is covering it. In the bottomsheet case, the dimview was always covering it
The alpha was always 0 in the closed state, but the player doesn't care about that
Setting dimView to INVISIBLE makes the player work. It will still pause when a sheet is opened, but that seems like expected behavior, and it can be resume after closing the sheet