Split transformer and dim alpha API, support composite transformers and remove default
Default was only there as a convenience so people didn't always have to pass a new transformer in when showing. This allows for greater flexibility for those that have multiple views/fragments that are supposed to interact with the content while cleaning up the API a bit. Next part will include a proper builder for this now that we have a slimmed down constructor setup